[FEAT] 음식 디테일 수정

+++ app/src/main/java/com/lee989898/todayeat/src/fooddetail/model/ResponseFoodDetail.kt
@@ -16,7 +16,7 @@ data class GetFoodRes(
 )
 
 data class MenuDto(
-    val locationX: Int,
-    val locationY: Int,
+    val locationX: Double,
+    val locationY: Double,
     val menuName: String
 )
